Pequeños cambios y optimizaciones
- Se ha comentado el ejemplo incluido para la mejor comprension.
- Pequeños cambios y optimizaciones gracias a Lucario448

http://forum.arduino.cc/index.php?topic=399485

+++ MillisTimer.h
@@ -5,12 +5,12 @@
 
 class MillisTimer {
 	public:
-		MillisTimer(int firstInterval);
+		MillisTimer(unsigned long firstInterval);
 		MillisTimer();
 		boolean checkTimer();
-		void setInterval(int newInterval);
+		void setInterval(unsigned long newInterval);
 		void setTimer();
-		int interval;
+		unsigned long interval;
 		unsigned long previousMillis;
 
 };
